Title This is about Lab1,Question B. It is about Writing a C++ program to read in an integer and print out the corresponding sequence of numbers.
Since we will do something different when the integer is odd and even,
so I use if else statement to make the condition different.
Since we want to repeat the if else statement,
so I use do while statement to make it repeat.

Hope that this is the right way to show the running script.
Otherwise, the only thing I can do is to write one more time.